code refactorinh

+++ src/main/java/com/shop/online/shopingMall/service/UserService.java
@@ -8,6 +8,7 @@ import com.shop.online.shopingMall.exception.NotFoundUserException;
 import com.shop.online.shopingMall.domain.User;
 import com.shop.online.shopingMall.dto.user.UserDto;
 import com.shop.online.shopingMall.dto.user.UserLoginResponseDto;
+import com.shop.online.shopingMall.exception.UnregisterFailExcpetion;
 import com.shop.online.shopingMall.repository.UserRepository;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
@@ -52,6 +53,10 @@ public class UserService {
         return userRepository.save(new User(userDto, encodePassWord));
     }
 
+
+    /**
+    * 해당 이메일이 존재하는지 확인을 한다.
+    * */
     public boolean checkEmail(String email) {
         return userRepository.existsByEmail(email);
     }
@@ -63,7 +68,7 @@ public class UserService {
     * 2) 배송중인 물품이 없을 경우 유저 상태를 DELETE로 변경하고, 카드 등록키와 카드 상태도 DELETE로 변경
     * */
     @Transactional(rollbackFor = Exception.class)
-    public void unRegister(Long id) throws NotFoundUserException {
+    public void unRegister(Long id) {
         User user = userRepository.findById(id).orElseThrow(NotFoundUserException::new);
         isDeletableUser(user);
     }
@@ -76,13 +81,21 @@ public class UserService {
         return user;
     }
 
+    /**
+    * 해당 유저가 현재 주문하고 있는 상태일 경우
+    * UnregisterFailExcpetion 호출
+    *
+    * 아닐 시
+    * 유저의 상태를 삭제로 변경하고, 갖고있는 카드정보와 카드 상태를 삭제로 변경한다.
+    * */
     private void isDeletableUser(User user) {
-        for (Order order : user.getOrderList()) {
-            if (isOrderReadyPresent(order)) {
-                return;
-            }
+        boolean isExist = user.getOrderList().stream().anyMatch(this::isOrderReadyPresent);
+
+        if (!isExist) {
+            user.delete();
+        } else {
+            throw new UnregisterFailExcpetion("주문중인 제품이 있습니다.");
         }
-        user.delete();
     }
 
     private boolean isOrderReadyPresent(Order order) {
